adding repository information.

+++ core-services/src/main/java/org/springsource/examples/sawt/services/jdbc/Main.java
@@ -1,10 +1,8 @@
 package org.springsource.examples.sawt.services.jdbc;
 
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
+import org.apache.commons.logging.*;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springsource.examples.sawt.CustomerService;
-import org.springsource.examples.sawt.services.jpa.JpaConfiguration;
 import org.springsource.examples.sawt.services.model.Customer;
 
 public class Main {
@@ -22,12 +20,12 @@ public class Main {
 
         Customer customer = customerService.createCustomer("Josh", "Long");
 
-        Customer retrievedCustomer = customerService.getCustomerById(customer.getId());
+        Customer retrievedCustomer = customerService.getCustomerById(customer.getId().longValue());
 
         log.info(String.format("customer.id (%s) == retrievedCustomer.id (%s)?  %s",
                 customer.getId(), retrievedCustomer.getId(), customer.getId().equals(retrievedCustomer.getId())));
 
-        Customer updatedCustomer = customerService.updateCustomer(customer.getId(), "Jdbc", "Lover");
+        Customer updatedCustomer = customerService.updateCustomer(customer.getId().longValue(), "Jdbc", "Lover");
         log.info(String.format("updated customer's firstName: %s", updatedCustomer.getFirstName()));
 
     }
